Added more types in src/types/Chat.ts
