+++ src/components/chat/w2wChat/chatBox/ChatBox.tsx
@@ -25,7 +25,7 @@ import * as w2wHelper from 'helpers/w2w/';
 import { generateKeyPair } from 'helpers/w2w/pgp';
 import useToast from 'hooks/useToast';
 import { useResolveEns } from 'hooks/useResolveEns';
-import { AppContext, Context } from 'sections/chat/ChatMainSection';
+import {  Context } from 'sections/chat/ChatMainSection';
 import HandwaveIcon from '../../../../assets/chat/handwave.svg';
 import { caip10ToWallet, decryptAndVerifySignature, encryptAndSign, walletToCAIP10 } from '../../../../helpers/w2w';
 import { fetchInbox, fetchIntent } from 'helpers/w2w/ipfs';
@@ -35,7 +35,7 @@ import Lock from '../../../../assets/Lock.png'
 import LockSlash from '../../../../assets/LockSlash.png'
 
 //Importing Interfaces
-import { ConnectedUser, Feeds, MessageIPFS, MessageIPFSWithCID, User } from 'types/Chat';
+import { AppContext, ConnectedUser, Feeds, MessageIPFS, MessageIPFSWithCID, User } from 'types/Chat';
 
 // Internal Configs
 import { appConfig } from 'config';
